Testing the vibe service layer

+++ src/main/java/com/ReVibe/service/VibeService.java
@@ -24,6 +24,7 @@ public class VibeService {
     
     public Vibe saveVibe(Vibe vibe){
         log.info("Saving a new vibe");
+        vibe.setVibeLike(0);
         Vibe newVibe = vibeRepository.save(vibe);
         log.info("{} saved", newVibe);
     	return newVibe;
@@ -48,17 +49,24 @@ public class VibeService {
     }
 
     public Like like(int vibeId, int accountId) {
-        log.info("Account {} toggling like for vibe {}", accountId, vibeId);
-        Like like = likeRepository.findByVibeIdAndUserId(vibeId, accountId);
-        
-        if (like == null){
-            log.info("Account {} adding like to vibe {}", accountId, vibeId);
-            return likeRepository.save(new Like(vibeId, accountId));
-        } else{
-            log.info("Account {} removing like from vibe {}", accountId, vibeId);
-            likeRepository.delete(like);
-            return null;
+        Vibe vibe = vibeRepository.findById(vibeId).get();
+        if (vibe != null){
+            log.info("Account {} toggling like for vibe {}", accountId, vibeId);
+            Like like = likeRepository.findByVibeIdAndUserId(vibeId, accountId);
+
+            if (like == null){
+                log.info("Account {} adding like to vibe {}", accountId, vibeId);
+                vibe.setVibeLike(vibe.getVibeLike()+1);
+                vibeRepository.save(vibe);
+                return likeRepository.save(new Like(vibeId, accountId));
+            } else{
+                log.info("Account {} removing like from vibe {}", accountId, vibeId);
+                vibe.setVibeLike(vibe.getVibeLike()-1);
+                vibeRepository.save(vibe);
+                likeRepository.delete(like);
+            }
         }
+        return null;
     }
     
     public List<Like> findLikesByVibeId(int vibeId){
